Performance pass!
1. Adopt & fork react-wastage-monitor (now react-frugal) to put
useless rerenders on blast. Enabled with ITCH_REACT_FRUGAL=1
2. Add launch configuration for it ("Review purity")
3. Release new version of ftl-redux-electron-store that does
not {...state} the state object if .updated and .empty are empty,
and that always dispatches actions to the browser even if they
are empty (for browser-side reactors or logging)
4. Add lru-memoize dependency (fastest memoizer in town), use it for
abstractions that are handy yet cpu-consuming, like space or
get-game-status. Also custom typings for it.
5. Introduce a `tick` action, dispatched every second, for #1481
6. For our HOCs (higher-order components), set the displayName properly