Closes #1510, 'build' field in cave model
In longer form:
- Redesign 'revert to' modal completely, group by month, format
dates in a smarter way, make sure user version is shown, remove
some text, turn entries into buttons and remove the manual build ID
input field
- Change wording from 'Revert to build X' to 'Switch to a different
version' in an attempt to be more user-friendly.
- Add a ctrl+shift+click inspector modal for download row
- A `build` JSONField has been added to the cave model, it
completely replaces & deprecates the `buildId` and `buildUserVersion`
columns. A migration has been written that converts the old
format to the new format. See #1534 for a future plan to clean up
database schemas
- In general, be more consistent about how we work with builds
throughout the codebase

Also not in short message:
- Fix global game update check (broken by sqlite-based rewrite, it
was just iterating an empty table of caves..)
- Add the `morphing` field to the caves model and use it to mark
whenever we were doing a 'revert/heal/upgrade' operation
- On launch, when `morphing` is set, force a heal and show a
notification saying what happened. Note that the item doesn't open
automatically after, the jury's out on what the best UX is for that.

Various:
- Introduce `formatBuildVersion` helper (should be in format, erm).
Also, that helper is not used by the revert-cave dialog, it has its
own graphical-oriented formatting.
- Pass 'db' to the preboot reactor via an argument instead of importing
it directly, to be consistent with all other reactors