Remove dependency on webContents.history
The history field of Electron's webContents module is no longer present
beyond Electron 10, so we have to stop using it in order to upgrade.

webContents.history is ostensibly used to manage a subset of browsing
history. My observation is that the global history (which includes
itch:// URLs) is a superset of the webContents.history, so the theory is
that we could rely entirely on the global history and back-forward
navigation would work just as well.

This commit introduces a new bug: after entering a URL such as
`http://itch.io` into the address bar, the app will change the URL to
`https://itch.io/` (note the change in protocol and the added trailing
slash), and the history will end up with two entries when there should
be only one. This means that going back from `https://itch.io/` will
navigate to `http://itch.io` only to loop you back to
`https://itch.io/`, rendering the back button useless in this particular
scenario. I can't figure out how to fix this correctly, so to work
around this in the common case, I'm changing the URL of the Explore
sidebar entry so that at least the itch.io homepage won't have this
problem.

Closes #2598