Convert butler callback system to structured message
Butler's `call` operation accepts a callback for the conversation,
in order to listen to particular events. This callback system is
challenging with context isolation, since movement between contexts
strips the object of its prototype. The whole purpose of the
callback system is to mutate the Conversation object.

Instead, we can pass a series of data messages that describe the
operation we'd like to happen in the main process, rather than
passing the conversation to the renderer process and trying to
mutate it there (which wouldn't work anyway).