Dispatch events to watchers hackily
After lots of fighting with store enhancers, I've given up, and
hacked the action routing into the reducer. There's certainly a
proper redux way of doing this, but at this stage I care more about
getting the app functioning smoothly.

This is comparable to the postDispatchCallback in the old
`redux-electron-store`