Continuously communicate to main when renderer is ready
Running locally, it's almost always the case that the renderer
process sends the `rootWindowReady` event before the main process
has started listening for it. Since we only send that event once,
this causes the main process to get stuck in a waiting state.

By sending it on effectively every tick, we're sure that the main
process, if in a good state, will handle it. As written, the event
handler will be a no-op if it has already acted on the event
successfully