Fix handling of new windows from browser
This resolves two bugs:

1. Links with `target="_blank"` not opening when left-clicked
2. Non-itch urls opening in the app browser (rather than external
   browser) when ctrl-clicked

The second was occurring because the wrong `loadURL` function was
being called, and we were bypassing our itch.io domain restriction.
The first was due to us not opting in to allowing popups, which
appears to have become required in Electron 22 (not confirmed, but
that was the version where window handling logic most significantly
changed).