Fix captcha handling
Captcha resolution wasn't IPC compatible. It depended on the
renderer process having direct access to `window` in a way that
hasn't been true since we upgraded to Electron 12.

The current event flow for captcha resolution is a bit roundabout,
in that it moves from:

contained webview -> main process -> renderer process

This could conceivably be shortened, but I don't know that its
necessarily valuable to pursue